+++ templates/watchlist_me.html
@@ -5,7 +5,7 @@
 <div class = "container">
   <h1>Welcome, {{ session['username'] }}</h1>
   <p>Soon, you will be able to view your watchlists on this page</p>
-  <a class="btn btn-primary btn-lg" href="{{ url_for('watchlist_new_list') }}" role="button">New List</a>
+  <a class="btn btn-primary btn-lg" href="{{ url_for('watchlist_blueprint.watchlist_new_list') }}" role="button">New List</a>
 
 </div>
