+++ templates/watchlist_mylists.html
@@ -12,7 +12,7 @@
 		  <div class="card-body">
 		    <h5 class="card-title">{{ l[3] }}</h5>
 		    <p class="card-text text-muted">Created: {{ l[4] }}</p>
-		    <a href="{{ url_for('watchlist_list_view', lid = l[0]) }}" class="btn btn-success">View</a>
+		    <a href="{{ url_for('watchlist_blueprint.watchlist_list_view', lid = l[0]) }}" class="btn btn-success">View</a>
 		  </div>
 		</div>
 		<br>
@@ -21,7 +21,7 @@
 	<p>No lists found, create a list</p>
 	{% endif %}
 	<br>
-	<a class="btn btn-primary btn-lg" href="{{ url_for('watchlist_new_list') }}" role="button">New List</a>
+	<a class="btn btn-primary btn-lg" href="{{ url_for('watchlist_blueprint.watchlist_new_list') }}" role="button">New List</a>
 
 </div>
